requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
asyncer==0.0.2
attrs==22.2.0
black==23.3.0
boto3==1.26.104
botocore==1.29.104
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
exceptiongroup==1.1.1
fastapi==0.95.0
filetype==1.2.0
flake8==6.0.0
flatbuffers==23.3.3
frozenlist==1.3.3
h11==0.14.0
humanfriendly==10.0
idna==3.4
ImageHash==4.3.1
imageio==2.27.0
iniconfig==2.0.0
jmespath==1.0.1
lazy_loader==0.2
llvmlite==0.39.1
mccabe==0.7.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
onnxruntime==1.14.1
opencv-python-headless==4.7.0.72
packaging==23.0
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.2.0
pluggy==1.0.0
pooch==1.7.0
protobuf==4.22.1
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
PyMatting==1.1.8
pytest==7.2.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyWavelets==1.4.1
rembg==2.0.32
requests==2.28.2
s3transfer==0.6.0
scikit-image==0.20.0
scipy==1.10.1
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
sympy==1.11.1
tifffile==2023.3.21
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
uvicorn==0.21.1
watchdog==3.0.0
yarl==1.8.2
